requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.4
exceptiongroup==1.2.2
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-cov==6.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.25
tomli==2.0.2
typing_extensions==4.9.0
urllib3==2.2.3
Werkzeug==3.0.1
zipp==3.20.2
